+++ src/Common/Socket.h
@@ -196,8 +196,8 @@ namespace dcclite
 			bool Send(const NetworkAddress &destination, const void *data, size_t size);
 			std::tuple<Status, size_t> Send(const void *data, size_t size);
 
-			std::tuple<Status, size_t> Receive(NetworkAddress &sender, void *data, const size_t size, const bool truncate = false);
-			std::tuple<Status, size_t> Receive(void *data, size_t size);
+			std::tuple<Status, int> Receive(NetworkAddress &sender, void *data, const int size, const bool truncate = false);
+			std::tuple<Status, int> Receive(void *data, int size);
 
 			bool JoinMulticastGroup(const IpAddress &address);
